Update patch set 5
Patch Set 5:

> The Attention Set's "one of you three reviewers are expected to review this change" opens up for an implicit negotiation of who that "one of you" should be and the answer is surprisingly often "someone else".

ben made a point of it meaning "all of you", by definition.
however, that's just not how reality works - that definition simply mutates "who" to "who first", because we know quite well that the first reviewer most likely has the most work. also, if everyone jumped on it, there would be a high risk of duplicating work (btw, i filed a feature request for real-time activity indication a while ago).

i think your other points boil down to the same conclusion i came to: the stated goal of replacing human judgement and communication with complicated rules is simply unrealistic. to make it work reliably, people would have to strictly adhere to micro-processes (like actually resolving every thread even in small reviews), and that simply doesn't match real-world behavior.

so imo it's better to err on the side of verbosity, and make sure that the information necessary to quickly decide whether action is expected from a particular user is prominently presented.

Patch-set: 5