Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > The Attention Set's "one of you three reviewers are expected to review this change" opens up for an implicit negotiation of who that "one of you" should be and the answer is surprisingly often "someone else".
> 
> ben made a point of it meaning "all of you", by definition.
> however, that's just not how reality works - that definition simply mutates "who" to "who first", because we know quite well that the first reviewer most likely has the most work.

As I think you are alluding to, I was not trying to convey the intended meaning of assignee/AS, I was describing how I see that they will most commonly be interpreted. Regardless of attempts to define them otherwise.

> i think your other points boil down to the same conclusion i came to: the stated goal of replacing human judgement and communication with complicated rules is simply unrealistic. to make it work reliably, people would have to strictly adhere to micro-processes (like actually resolving every thread even in small reviews), and that simply doesn't match real-world behavior.
> 
> so imo it's better to err on the side of verbosity, and make sure that the information necessary to quickly decide whether action is expected from a particular user is prominently presented.

If you by that mean that the intent and/or expectations of the change-owner f.i. should be clearly decipherable, as it is with assignee today.

Having time to think about it I look at 3 different sets of reviewers (term used loosely here).

A. User that needs to take action.
B. Users of which at least one should review the change.
C. Users that should be aware of the change happening so that they may take action/review if they feel that it is necessary.

Which maps somewhat to current A - assignee, B - reviewer, C - CC.
The problem with the current behavior are obvious to all, f.i. changes are marked with "needs action" (bold in change list) after any action on the change.

I feel that AS should be able to cover A & B (I obviously believe that it doesn't cover 'A' ATM).

A - reviewer in AS + flag/label, B - reviewer in AS, C - reviewer not in AS

Patch-set: 5
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>