Update patch set 5
Patch Set 5:

> Having time to think about it I look at 3 different sets of reviewers (term used loosely here).
> 
> A. User that needs to take action.
> B. Users of which at least one should review the change.
> C. Users that should be aware of the change happening so that they may take action/review if they feel that it is necessary.
> 
> Which maps somewhat to current A - assignee, B - reviewer, C - CC.
> The problem with the current behavior are obvious to all, f.i. changes are marked with "needs action" (bold in change list) after any action on the change.
> 
> I feel that AS should be able to cover A & B (I obviously believe that it doesn't cover 'A' ATM).
> 
> A - reviewer in AS + flag/label, B - reviewer in AS, C - reviewer not in AS

C should still be covered by CC, before and after this proposal.

Don't think of the AS as a different set of people. It is just a boolean attached to owner and reviewers (and sometimes CC, but only in rare cases).

I think we are talking past each other, so let's come up with some concrete workflows. I will use -> to denote who is in the AS and needs thus to take action on a change. When you send something for review to multiple reviewers, then IMO you would always want:

     owner
  -> reviewer A
  -> reviewer B
  -> reviewer C

I cannot imagine a situation where I want to send a change to multiple reviewers, but only request action from one specifically. Do you agree or have an example for me?

If you think that A, B and C are likely to duck away from being responsible, then you should maybe start with (or fix your workflow for findinging the correct reviewer):

     owner
  -> reviewer A
  CC person B
  CC person C

While I personally don't see a need for it, you are still able to start a review with multiple reviewers and just a single reviewer assigned, if you want to encourage a review from B and C slightly more than just by CCing them:

     owner
  -> reviewer A
     reviewer B
     reviewer C

That would also cover the three sets of reviewers that you are looking for (sorry for also using A,B,C :-/)

  A: A
  B: A, B, C
  C: CC

I want to repeat again here that the proposal is just about extending the assignee feature by two things:

  * Allowing multiple users to be assigned (but you can still assign to just one, if you want to).
  * Automatically adding and removing assignments when a change message is published.

Starting from my initial example, if reviewer B thinks that they don't have time to do the review, then they would unassign themselves with a "don't have the bandwidth at the moment" and change to 

     owner
  -> reviewer A
     reviewer B
  -> reviewer C

Patch-set: 5