Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> > Having time to think about it I look at 3 different sets of reviewers (term used loosely here).
> > 
> > A. User that needs to take action.
> > B. Users of which at least one should review the change.
> > C. Users that should be aware of the change happening so that they may take action/review if they feel that it is necessary.
> > 
> > Which maps somewhat to current A - assignee, B - reviewer, C - CC.
> > The problem with the current behavior are obvious to all, f.i. changes are marked with "needs action" (bold in change list) after any action on the change.
> > 
> > I feel that AS should be able to cover A & B (I obviously believe that it doesn't cover 'A' ATM).
> > 
> > A - reviewer in AS + flag/label, B - reviewer in AS, C - reviewer not in AS
> 
> C should still be covered by CC, before and after this proposal.
> 
> Don't think of the AS as a different set of people. It is just a boolean attached to owner and reviewers (and sometimes CC, but only in rare cases).

IIUC it's clearly more than a boolean since it has a data-structure.
If it is just a boolean on the reviewer object perhaps it should be renamed "attention-flag".

I view it as

    Attention-set is a sub-set of reviewers.

My suggested approach (I use the term assignee but I don't mean the current behavior where assignee is a state on the change but rather a state of the attention-set).

    assignee is a sub-set of attention-set is a subset of reviewers
> 
> I think we are talking past each other, so let's come up with some concrete workflows. I will use -> to denote who is in the AS and needs thus to take action on a change. When you send something for review to multiple reviewers, then IMO you would always want:
> 
>      owner
>   -> reviewer A
>   -> reviewer B
>   -> reviewer C
> 
> I cannot imagine a situation where I want to send a change to multiple reviewers, but only request action from one specifically. Do you agree or have an example for me?

My point is that you request action from those you add as reviewers but you demand action from the one you have set as assignee. In the assignee case a viable action is to just remove one-self as assignee, showing that you cannot do a review ATM. Which is valuable information that the owner can act upon.

> 
> If you think that A, B and C are likely to duck away from being responsible, then you should maybe start with (or fix your workflow for findinging the correct reviewer):
> 
>      owner
>   -> reviewer A
>   CC person B
>   CC person C
> 
> While I personally don't see a need for it, you are still able to start a review with multiple reviewers and just a single reviewer assigned, if you want to encourage a review from B and C slightly more than just by CCing them:
> 
>      owner
>   -> reviewer A
>      reviewer B
>      reviewer C
> 
> That would also cover the three sets of reviewers that you are looking for (sorry for also using A,B,C :-/)

As I have stated before this approach is very fragile since anyone can add reveiwer D, and then reviewer D gets added to the attention-set. I have to notice this, go back to the change and remove reviewer D from the attention-set to make my expectations clear again.

> 
>   A: A
>   B: A, B, C
>   C: CC
> 
> I want to repeat again here that the proposal is just about extending the assignee feature by two things:
> 
>   * Allowing multiple users to be assigned (but you can still assign to just one, if you want to).

As I have stated before, having assignee as a set instead of a singleton defeats the purpose of assignee, shared responsibility needs to be negotiated. Shared responsibility without negotiation is IMO in reality often the same as no responsibility.
As per my comment above, one single member of the attention-set is a very fragile state.

>   * Automatically adding and removing assignments when a change message is published.
> 
> Starting from my initial example, if reviewer B thinks that they don't have time to do the review, then they would unassign themselves with a "don't have the bandwidth at the moment" and change to 
> 
>      owner
>   -> reviewer A
>      reviewer B
>   -> reviewer C

Don't know if I've mentioned or made this clear before, but this is what I understand from the proposal.

1. AS will replace the reviewed flag, in the sense that the reviewed flag was meant as reviewed = false => needs action.
2. Unlike reviewed flag AS will have configurable state-transitions [IMO this is where the awsomeness of this proposition lies].
But then it goes:
3. Since "needs action" is a form of assignment, and AS will be better at deciding if the change needs action from a user than the former reviewed-flag there's no longer the need to assign a single user to a change.

I believe that the assumption of 3. is flawed and insufficient reasons to remove the functionality, hence my initial CR-1.

Viable reasons could be:
"Having both the assignee-feature and AS would be unnecessary complex so we opt to remove the possibility to assign a change to a single user."
I.e. even though it's an appreciated and well-used functionality we feel that the benefits of keeping it are not worth the drawbacks from removing it.

One of the questions that AS is aimed to answer according to the design is "Who's turn is it?", and the functionality of assignee does that (whereas I obviously clearly believe that the current iteration of AS doesn't do that satisfactory).

Assignee:
owner -> Reviewer-A : I need you to look at this
Reviewer-A -> Reviewer-B: I don't have the time ATM but Reviewer-B has agreed to look at it.
Reviewer-B -> owner: I made some comments about thing that need to change.
owner - Reviewer-B: I made alterations as per your request.
...

And without a clear statement of whose turn it is users can easily draw the conclusion that it's not their turn.

Patch-set: 6
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>