Update patch set 6
Patch Set 6:

(1 comment)

> Let's find out where exactly you see the technology excess. [...]
> Can you comment individually on which things you consider too complex?

the whole thing. ;)
the state machine is supposed to have only two states, but the transitions have about a bazillion of conditions. i already pointed out why that is bad.

> - "Latest activity on the change by any other user" is a very poor indicator for bringing something back to my attention.

> I would first need to understand "what happened" and then decide, if this really requires action from my side.

yes, you would, and i say there is no reliable way around that.
therefore, an imo more constructive approach would be designing ways to present the events in a way that facilitates quick decision making. e.g., hovering each "unseen" item on the dashboard could show a tooltip with relevant events and a button to "mark as seen" right away.

> Uploaded patchsets or comments by other reviewers are just a distraction most of the time.

i can take that ad absurdum by pointing out that reviews as a whole are a distraction most of the time. ;)

but more seriously, the key is that "most of the time" is not sufficient grounds for dismissing the alternative. this isn't a facebook wall, where filtering content according to expected engagement is the goal (from fb's pov, anyway). and the machine cannot possibly have enough context to reliably decide which contributions by other users would warrant a followup from me. even this review already proved that point ...

> - "Last visited" as indicator for having taking action is also poor, because you may (A) look at a change and then decide to postpone the review and (B) you don't want to indicate to other users that you are done while you are still reviewing the change (or still addressing comments as the owner).

yes, this is a problem, indeed. so a) the timestamp must not be updated until you leave and b) when you leave without apparent action, it could ask you whether you intend to come back.

> - From a data model POV [...]

that's hardly the relevant pov here. ;)

> > some use cases you listed seem like mindless control freakery. ^^
> 
> Which ones?

just the buzzword bingo surrounding the word "manager". ^^

but more seriously, it's an overall impression.

i'm all for automating decisions, but only when it can be done reliably and doesn't require a straitjacket of otherwise unnecessary processes to make it so. otherwise, the automaton should only provide hints.

Patch-set: 6