Update patch set 6
Patch Set 6:

(27 comments)

I have substantially revised the proposal. Apart from addressing all comments I have removed the "rules and conditions" part (moved a lot of content to workflow.md). Instead I have added "Defaults", "Expectations" and "Snoozing" sections to workflow.md. This emphasizes that I am aiming for just some useful defaults instead of a complicated state machine with lots of automation.

PTAL


@Sven

> My point is that you request action from those you add as reviewers but you demand action from the one you have set as assignee.

Can you elaborate what the difference between "request action" and "demand action" is?

> As I have stated before this approach is very fragile since anyone can add reviewer D, and then reviewer D gets added to the attention-set.

I don't think that is an issue. Someone from the sideline adding a second user to the attention set without good reason has never come up as a problem in Critique (~30,000 users in ~4 years). If you feel that your review is only going to make good progress with just one user assigned, then as the change owner you have that under your control.

OTOH you have to acknowledge that there are changes that you genuinely want two reviewers to look at in parallel, either because you want a review from both of them, or because the two reviewers are supposed to focus on two separate aspects that do not depend on each other. In that case it is really useful to assign both reviewers, not?

Patch-set: 6