Update patch set 7
Patch Set 7:

(3 comments)

Have you done a survey to see whether users of f.i. *-review feel that being able to assign a single user to a change is a valuable feature or not?

Attention Set, I feel, is a good replacement for the "reviewed" flag.
However it cannot replace the assignee feature (if implemented according to the current revision of the design-document).

If I haven't been clear before:
I honestly feel that the reasons for removing the assignee feature, as stated in this design-document, is based on loose assumptions about what *some* users might think about the implementation of it and issues with the current implementation, most of which you could quite easily be fixed within the boundaries of the assignee feature.

If you by any means necessary want to get rid of the assignee feature there's little use in straw-manning it, instead just be honest and clear about the reasons whatever they may be:

* We don't think that the Gerrit users benefits from being able
  to assign a change to a single user so we will remove it.
* We prioritize Attention-Set and we believe it conflicts with
  assignee in such a way that we will remove it.
* We believe that being able to single out a user in an Attention-Set
  defeats the purpose of the Attention-Set feature so we will not make
  that feature a part of it.

For me it's quite apparent that Attention Set cannot replace assignee.
I believe that assignee is a feature that:
* Brings clarity to the question "who's turn is it".
* Enables reviewers to prioritize their reviews according to change owner expectations.
* Enable change owners to take action when a review stalls.

Changes were reviewed quite readily before the assignee feature was ever implemented, but I personally regard removing it as a degradation, Attention-Set or no Attention-Set.

> @Sven
> 
> > My point is that you request action from those you add as reviewers but you demand action from the one you have set as assignee.
> 
> Can you elaborate what the difference between "request action" and "demand action" is?
> 
> > As I have stated before this approach is very fragile since anyone can add reviewer D, and then reviewer D gets added to the attention-set.
> 
> I don't think that is an issue. Someone from the sideline adding a second user to the attention set without good reason has never come up as a problem in Critique (~30,000 users in ~4 years).

What do you mean? That no-one has complained about it in 4 years or that you have looked in the logs for the past 4 years and there have never been anyone other then the owner that have added users to an Attention Set?

> If you feel that your review is only going to make good progress with just one user assigned, then as the change owner you have that under your control.

Not if other users can add users to the Attention Set I don't.

> 
> OTOH you have to acknowledge that there are changes that you genuinely want two reviewers to look at in parallel, either because you want a review from both of them, or because the two reviewers are supposed to focus on two separate aspects that do not depend on each other. In that case it is really useful to assign both reviewers, not?

As I explained previously 99% of our changes has had at least 2 reviewers, it's an implicit requirement. I have also in more than one occasion been in the situation where I want two separate reviewers to review two separate part of a change. And I can honestly say that I wouldn't have been helped by being able to assign a change to two users simultaneously. As I believe I have explained previously "I don't believe that assigning something to two different users have the desired effect".

I don't use assignee for all my changes.
The most common use-case in the scenarios above is:
1. I upload a change.
2. 3-4 users are automatically added as reviewers.
3. I get a code-review from one user.
4. When so much time has passed without a second review that I feel that I must get things going (the amount of time is depending on how urgent the change is) I assign the change to one of the other reviewers.
5. Often I get a response fairly quickly either in the form of review or "I don't have the time right now".

In the second use-case once I have a review by one of the (using assignee or not) I would add the other one as assignee.
Another common use-case is that I get CR+1 and CR+2 but I want a third person to review it (f.i. because they have expertise in a certain area) then I would add that third person as assignee.

Can you honestly say that you would review something that was ready for submit just because you were in the Attention-Set?
I would guess that most users would interpret it as a mistake that they are still in the Attention Set and just remove themselves from it.

Patch-set: 7