Update patch set 8
Patch Set 8:

(3 comments)

@Oswald: I will reply to your comments separately.

First some meta comments:

I really appreciate your thorough reviews, but I think we reaching a limit of where this is an efficient use of our time. Maybe we could meet over VC?

It would be great, if you could write a separate solution.md file in the spirit of Gerrit's design doc process:

https://gerrit-review.googlesource.com/Documentation/dev-design-docs.html
"Anyone who has ideas for an alternative solution uploads a change with a solution-<n>.md that describes their solution."

Currently I do not yet have a good understanding of what the best solution is to the primary use cases that you prefer. Just keeping the assignee feature as is without any attention set? Having both features in parallel or just one configurable per host?

> Have you done a survey to see whether users of f.i. *-review feel that being able to assign a single user to a change is a valuable feature or not?

Using the attention set you *can* assign a single user.
 
> I honestly feel that the reasons for removing the assignee feature, as stated in this design-document, is based on loose assumptions about what *some* users might think about the implementation of it and issues with the current implementation, most of which you could quite easily be fixed within the boundaries of the assignee feature.

Can you propose a solution of how to fix the assignee feature?
 
> If you by any means necessary want to get rid of the assignee feature

My primary goal is not to get rid of the assignee feature: 1. I see the attention set as the same feature, just evolved a little bit. I would also be open to keeping the name and calling it "set of assignees" or something like that. 2. I would be open to allowing you to keep the assignee feature. It just seemed to me that using both at the same time does not make sense, and that attention set can be a super set of what assignee offers.

> * We don't think that the Gerrit users benefits from being able
>   to assign a change to a single user so we will remove it.

You can assign a change to a single user with the attention set.

> * We prioritize Attention-Set and we believe it conflicts with
>   assignee in such a way that we will remove it.

Yes, I do see a huge overlap. From my point of view both features answer the question "Whose turn is it?" and "Who has to take action?" If you disagree maybe you can write about that in your alternative solution.md.

> * We believe that being able to single out a user in an Attention-Set
>   defeats the purpose of the Attention-Set feature so we will not make
>   that feature a part of it.

If just one person is in the attention set, then this user is the only one who is responsible for taking action. From my point of view that is "singling out", so yes, having two grades of singling out and being responsible would muddy the waters IMO.
 
> > > My point is that you request action from those you add as reviewers but you demand action from the one you have set as assignee.
> > 
> > Can you elaborate what the difference between "request action" and "demand action" is?

Would still be interested in an answer. :-)

> > > As I have stated before this approach is very fragile since anyone can add reviewer D, and then reviewer D gets added to the attention-set.
> > 
> > I don't think that is an issue. Someone from the sideline adding a second user to the attention set without good reason has never come up as a problem in Critique (~30,000 users in ~4 years).
> 
> What do you mean? That no-one has complained about it in 4 years or that you have looked in the logs for the past 4 years and there have never been anyone other then the owner that have added users to an Attention Set?

No one has complained about that being a problem. From looking at the logs it does sometimes happen that reviewers add other reviewers to the attention set, but when they do, then they have good reasons for it. Imagine a reviewer in the US passing on a change to their colleague in the same time-zone, while the owner in Europe is still asleep.
 
> > If you feel that your review is only going to make good progress with just one user assigned, then as the change owner you have that under your control.
> 
> Not if other users can add users to the Attention Set I don't.

Why would they do that? Other users can also change the assignee. Has that been a problem?
 
> As I explained previously 99% of our changes has had at least 2 reviewers, it's an implicit requirement. I have also in more than one occasion been in the situation where I want two separate reviewers to review two separate part of a change. And I can honestly say that I wouldn't have been helped by being able to assign a change to two users simultaneously. As I believe I have explained previously "I don't believe that assigning something to two different users have the desired effect".
> 
> I don't use assignee for all my changes.

When you don't use "assignee" how do you know whose turn it is?

> Can you honestly say that you would review something that was ready for submit just because you were in the Attention-Set?

Yes, that is what all Google engineers do every day. If you are in the attention set, then it is your turn. You are expected to act. Being in the attention set at Google has the same strong meaning as being assigned at Axis. That is just a matter of convention (and the tool doing a good UX job at enforcing that convention, and Gerrit admins communicating that clearly to their users).

> I would guess that most users would interpret it as a mistake that they are still in the Attention Set and just remove themselves from it.

No, that is not what we experience every day. You also don't typically assign someone by mistake. The attention is not more error prone in that regard than the assignee feature.

Patch-set: 8