Update patch set 10
Patch Set 10:

An attempt to clarify my position:
* Attention-Set - is used for all changes. It's the norm.
* assignee      - is used when it's needed. It's the exception.

* Attention-Set - (like reviewer) is relationship that originates from a change.
You are in the AS *of* that change. You are *one-of-many*.
* assignee      - is a relationship that originates from a reviewer.
You are assigned *to* that change. You are *the-only-one*.

From the reviewers perspective:
You are added as reviewer on 20 changes.

There's no need for you to look at 12 of those ATM because the states of those changes haven't changed since you last reviewed them in such a way that there's a point in you revisiting them.
[Attention-Set describes this]

In one or two of remaining 8 changes you are the user that blocks progress.
[assignee describes this]

With assignee you have enough data to prioritize which of those 8 changes you should review first.

> Currently I do not yet have a good understanding of what the best solution is to the primary use cases that you prefer. Just keeping the assignee feature as is without any attention set? Having both features in parallel or just one configurable per host?

Attention-Set:
I think Attention-Set is a good match for doing what the "reviewed-flag" tries to accomplish, basically IMM "Is it time for me to look at this change again?".
But since you can adopt the logic/behavior it seems far superior to me.

So IMO Attention-Set should be the default feature to keep track of this and we can scrap the "reviewed-flag".

Assignee:
From previous discussion it should be obvious that some consider this a valuable and vital feature that cannot be replaced with AS whereas some consider it analogous to AS and merely a nuisance once Attention-Set is in place.

To cater to this I believe we should make assignee configurable on Gerrit instance level.
If assignee is enabled it pretty much has the same function as today.
If assignee is disabled it's removed from the UI and all REST endpoints replies with "Feature not enabled" or similar.
What you are left with is basically an empty field in the index, no?


> > > > My point is that you request action from those you add as reviewers but you demand action from the one you have set as assignee.
> > > 
> > > Can you elaborate what the difference between "request action" and "demand action" is?
> 
> Would still be interested in an answer. :-)
I have tried to answer this question before but to not leave you hanging I will try again.
UC1
I need reviews from (A|B|C)&D
I would "request action" from A,B and C. Since it's, on average, three times as likely to get a timely review from three people than from one.
I would "demand action" from D.

UC2
I need reviews from two of A,B,C and D.
I "request review" from A,B,C and D.
B reviews my change with a CR+1, after a day of no more reviews I "demand action" from one of the other reviewers to the change and get either a review or a NAK, if NAK I "demand action" from one of the other reviewers.

UC3
I need reviews from two of A,B,C and D.
I "request review" from A,B,C and D.
A reviews my change but gives it a CR-1.
I upload a new PS to fix the review comments and "demand action" from A.
I would be satisfied if one of [B,C,D] also reviews the change but I need for A to ack whether the changes I made was a sufficient answer to A's review-comments.

> 
> > > > As I have stated before this approach is very fragile since anyone can add reviewer D, and then reviewer D gets added to the attention-set.
> > > 
> > > I don't think that is an issue. Someone from the sideline adding a second user to the attention set without good reason has never come up as a problem in Critique (~30,000 users in ~4 years).
> > 
> > What do you mean? That no-one has complained about it in 4 years or that you have looked in the logs for the past 4 years and there have never been anyone other then the owner that have added users to an Attention Set?
> 
> No one has complained about that being a problem. From looking at the logs it does sometimes happen that reviewers add other reviewers to the attention set, but when they do, then they have good reasons for it. Imagine a reviewer in the US passing on a change to their colleague in the same time-zone, while the owner in Europe is still asleep.

Yes, that's a good reason to add someone to the AS which is exactly my point, we tried to use AS for something it wasn't intended for and there are now >1 reviewers in the AS so the first reviewer is no longer singled out.
This might seem like a petty difference but psychologically it makes a world of difference for that first reviewer in the AS.

>  
> > > If you feel that your review is only going to make good progress with just one user assigned, then as the change owner you have that under your control.
> > 
> > Not if other users can add users to the Attention Set I don't.
> 
> Why would they do that? Other users can also change the assignee. Has that been a problem?

You acknowledged above that it's quite reasonable to do so.
The problem is that assignee only means one thing. If you change the assignee you change the assignee.
If you don't know that there was only one reviewer in the AS to single that user out you cannot instinctively know that you defeat that purpose by adding another one.

>  
> > As I explained previously 99% of our changes has had at least 2 reviewers, it's an implicit requirement. I have also in more than one occasion been in the situation where I want two separate reviewers to review two separate part of a change. And I can honestly say that I wouldn't have been helped by being able to assign a change to two users simultaneously. As I believe I have explained previously "I don't believe that assigning something to two different users have the desired effect".
> > 
> > I don't use assignee for all my changes.
> 
> When you don't use "assignee" how do you know whose turn it is?

Generally you get the feedback you need in a timely manner from the people that are added as reviewers.
At times when you don't, you use assignee to make it obvious to the reviewer that you expect them to prioritize that particular change.

> 
> > Can you honestly say that you would review something that was ready for submit just because you were in the Attention-Set?
> 
> Yes, that is what all Google engineers do every day. If you are in the attention set, then it is your turn. You are expected to act. Being in the attention set at Google has the same strong meaning as being assigned at Axis. That is just a matter of convention (and the tool doing a good UX job at enforcing that convention, and Gerrit admins communicating that clearly to their users).

> > I would guess that most users would interpret it as a mistake that they are still in the Attention Set and just remove themselves from it.
> 
> No, that is not what we experience every day. You also don't typically assign someone by mistake. The attention is not more error prone in that regard than the assignee feature.

I concede that given the difference between assignee and AS, my example was quite poor.

Patch-set: 10
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>