Update patch set 10
Patch Set 10:

Here's the assignee usage/user for the past 30 days on our Gerrit instance with around 700-800 active users.
Around 10% uses it each month, which is as expected since it's only used when needed.
The set of users vary for each individual 30 day period as well as their usage frequency.
Over a 6 month period more than 100 users have used the assignee feature.

These are the users that used it and how often they've used it the past 30 days:

https://imgur.com/a/JaTfByX

We have not advertised the feature other than a notice that the feature is available when we upgraded to a version that supported it. It was adopted from the beginning has grown organically.

Patch-set: 10
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>