Update patch set 10
Patch Set 10:

Let me begin by clarifying that I have never stated, or suggested, that adding more reviewers leads to faster review-times, I have repeatedly made the reverse statement.

> Patch Set 10:
> 
> > UC1
> > I need reviews from (A|B|C)&D
> > I would "request action" from A,B and C. Since it's, on average, three times as likely to get a timely review from three people than from one.
> > I would "demand action" from D.
> 
> Sven, fyi, we did some research on our Gerrit logs and the results suggest that the behavior you describe here is actually associated with reviews that take longer (+5 hours time to submit for the median change having vs. not having at least one reviewer invited that did not end up voting).

These numbers are describing the fact that if a change doesn't get a timely review the owner is likely to invite additional reviewers.

A simple and, IMO, not uncommon use-case:
You upload a change and add reviewer A.
Reviewer A is otherwise engaged and after 5 hours without activity you add reviewer B.
 
> 
> On top of that, our data also suggests that using an assignee is not improving on that situation either. Changes having an assignee added at least once to a change that has at least one non-voting reviewer invited initially actually show increased time to submit by another 5.4 hours. (changes having only voting reviewers invited: 14.0 hours to submit; changes having at least one non-voting reviewer initially invited: 19.1 hours to submit; changes having a non-voting reviewer initially invited *and* at least one assignee added at some point: 24.5 hours to submit) 

I don't find these numbers surprising. To me they describe the fact that you are likely to add an assignee when you don't get a timely review. Which is one of the utilities of it.

> Of course, there is a caveat to this data, as we cannot infer causality given the changes in each category might be systematically different from each other as this was not an experimentally controlled situation. We don't have any evidence for significant differences between these buckets, though, and the changes having non-voting reviewers invited initially actually almost make up half of the changes we analyzed. However, we can see that using extra reviewers and assignee is at least not improving the situation for these changes.

> 
> So, overall, we believe that the assumption you are making in your UC1 of getting faster reviews this way might not hold, at least not based on the data we have from Google's Gerrit instances.

Depending if you (as you seem to do) infer that the effect of adding multiple reviewers or use assignee causes longer review-times OR if you (as I) infer that longer review-times causes the change-owner to add more reviewers or use assignee.

Let me also remind you that the fact that "increasing the number of users with responsibility tends to be interpreted by humans as if their individual responsibility has decreased" is the main argument for the assignee feature.

Patch-set: 10
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>